change a type.
git-svn-id: svn+ssh://gforge/svn/siconos/trunk@307 bf01a9d7-0df3-0310-8ad4-b1c94d9ae2d5

+++ Numerics/src/solverpack/src/dfc_2D_solver.c
@@ -81,7 +81,7 @@ int dfc_2D_solver(double *K1, double *F1, int *n, method *pt, double *U2 , doubl
   int           i, info, it_end;
   int           dim_q, dim_MM;
   int           iparamLCP[5];
-  int           dparamLCP[5];
+  double        dparamLCP[5];
 
   double        res;
   double        *MM, *q, *z, *w;
